requirements.txt
# Framework + server
fastapi>=0.115.0
uvicorn>=0.30.0

# Templates for the web UI
jinja2>=3.1.4

# HTTP client for calling OpenAI (embeddings + chat/completions)
httpx>=0.27.0

# Data validation and settings
pydantic>=2.7.0
python-dotenv>=1.0.1

# Math and vectors
numpy>=2.1.0

# PDF parsing
pypdf>=4.2.0
pdfminer.six>=20231228
pillow>=10.4.0
